+++ kontiki/testing/manager.py
@@ -140,6 +140,7 @@ class MockServiceManager:
         self,
         service_name,
         method_name,
+        /,
         *args,
         extra_headers=None,
         flow_id=None,